+++ flask_unchained/bundles/controller/resource.py
@@ -1,11 +1,13 @@
 from flask_unchained.string_utils import pluralize
-from py_meta_utils import deep_getattr
+from py_meta_utils import McsArgs, MetaOption, deep_getattr, _missing
 
 from .attr_constants import CONTROLLER_ROUTES_ATTR, REMOVE_SUFFIXES_ATTR
 from .constants import ALL_METHODS, INDEX_METHODS
 from .constants import CREATE, DELETE, GET, LIST, PATCH, PUT
-from .controller import Controller, ControllerMeta, _get_remove_suffixes
-from .utils import controller_name
+from .controller import (Controller, ControllerMeta, ControllerMetaOptionsFactory,
+                         UrlPrefixMetaOption as ControllerUrlPrefixMetaOption,
+                         _get_remove_suffixes)
+from .utils import controller_name, get_param_tuples
 
 
 RESOURCE_REMOVE_EXTRA_SUFFIXES = ['MethodView']
@@ -36,7 +38,7 @@ class ResourceMeta(ControllerMeta):
                 rule = '/'
             else:
                 route._is_member_method = True
-                route._member_param = deep_getattr(clsdict, bases, 'member_param')
+                route._member_param = cls._meta.member_param
             route.rule = rule
             controller_routes[method_name] = [route]
         setattr(cls, CONTROLLER_ROUTES_ATTR, controller_routes)
@@ -44,9 +46,69 @@ class ResourceMeta(ControllerMeta):
         return cls
 
 
-class UrlPrefixDescriptor:
-    def __get__(self, instance, cls):
-        return '/' + pluralize(controller_name(cls).replace('_', '-'))
+class UrlPrefixMetaOption(MetaOption):
+    def __init__(self):
+        super().__init__('url_prefix', default=_missing, inherit=False)
+
+    def get_value(self, meta, base_classes_meta, mcs_args: McsArgs):
+        value = super().get_value(meta, base_classes_meta, mcs_args)
+        if value is not _missing:
+            return value
+
+        prefix = controller_name(mcs_args.name,
+                                 deep_getattr(mcs_args.clsdict,
+                                              mcs_args.bases,
+                                              REMOVE_SUFFIXES_ATTR))
+        return '/' + pluralize(prefix.replace('_', '-'))
+
+    def check_value(self, value, mcs_args: McsArgs):
+        if not value:
+            return
+
+        assert isinstance(value, str), \
+            f'The {self.name} meta option must be a string'
+
+
+class MemberParamMetaOption(MetaOption):
+    def __init__(self):
+        super().__init__('member_param', default=_missing, inherit=True)
+
+    def get_value(self, meta, base_classes_meta, mcs_args: McsArgs):
+        value = super().get_value(meta, base_classes_meta, mcs_args)
+        if value is not _missing:
+            return value
+
+        return '<int:id>'
+
+    def check_value(self, value, mcs_args: McsArgs):
+        if mcs_args.meta.abstract:
+            return
+
+        assert isinstance(value, str) and len(get_param_tuples(value)) == 1, \
+            f'The {self.name} meta option must be in the format of ' \
+            f'<werkzeug_type:param_name>'
+
+
+class UniqueMemberParamMetaOption(MetaOption):
+    def __init__(self):
+        super().__init__('unique_member_param', default=None, inherit=False)
+
+    def check_value(self, value, mcs_args: McsArgs):
+        if mcs_args.meta.abstract or not value:
+            return
+
+        assert isinstance(value, str) and len(get_param_tuples(value)) == 1, \
+            f'The {self.name} meta option must be in the format of ' \
+            f'<werkzeug_type:param_name>'
+
+
+class ResourceMetaOptionsFactory(ControllerMetaOptionsFactory):
+    options = [option for option in ControllerMetaOptionsFactory.options
+               if not issubclass(option, ControllerUrlPrefixMetaOption)] + [
+        UrlPrefixMetaOption,
+        MemberParamMetaOption,
+        UniqueMemberParamMetaOption,
+    ]
 
 
 class Resource(Controller, metaclass=ResourceMeta):
@@ -88,7 +150,8 @@ class Resource(Controller, metaclass=ResourceMeta):
 
 
         class UserResource(Resource):
-            member_param: '<string:username>'
+            class Meta:
+                member_param: '<string:username>'
 
             def __init__(self, user_manager: UserManager = injectable):
                 super().__init__()
@@ -138,26 +201,10 @@ class Resource(Controller, metaclass=ResourceMeta):
     See also :class:`~flask_unchained.bundles.api.model_resource.ModelResource` from
     the API bundle.
     """
+    _meta_options_factory_class = ResourceMetaOptionsFactory
+
     class Meta:
         abstract = True
-
-    member_param = '<int:id>'
-    """
-    The URL parameter to use for member methods (get, patch, put, and delete).
-    """
-
-    unique_member_param = None
-    """
-    The URL parameter to use for member methods (get, patch, put, and delete)
-    when there is a conflict with a subresource's member_param.
-    """
-
-    url_prefix = UrlPrefixDescriptor()
-    """
-    The URL prefix to use for this Resource. Defaults to the pluralized,
-    kebab-cased controller name (eg UserController => '/users',
-                                    FooBarController => '/foo-bars')
-    """
 
     @classmethod
     def method_as_view(cls, method_name, *class_args, **class_kwargs):