+++ tests/bundles/controller/fixtures/bp_views.py
@@ -101,7 +101,8 @@ class RoleResource(Resource):
 
 
 class AnotherResource(Resource):
-    url_prefix = 'another'
+    class Meta:
+        url_prefix = 'another'
 
     def list(self):
         return self.render('list')