+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * fix the resource url prefix descriptor to convert to kebab-case instead of snake-case
 * rename `Controller.template_folder` to `Controller.template_folder_name`
 * add `Controller.make_response` as an alias for `flask.make_response`
+* convert attributes on `Controller`, `Resource`, and `ModelResource` to be `class Meta` options
 
 ## 0.5.1 (2018/07/25)